Add files via upload
dataToMGdb.py: dataToMGdb: Reads the data of the three sensors in the jsonl file and stores it in the local MongoDB. Note that the path needs to be adjusted.

visualization: Read the content generation chart for the specified mgdb, monitor for changes, and update the chart. Note that the path needs to be adjusted.